Restructure project (#1)
* Restructured to new project layout + changed dep mngmt to go mod
* update dependencies
* changed travis cfg
* introducing gihub actions (for future use)
