+++ .github/workflows/build.yml
@@ -1,0 +1,34 @@
+name: Build
+
+on:
+  push:
+    branches: [ "master", "main" ]
+  pull_request:
+    types: [opened, reopened, synchronize]
+    branches: [ "master", "main" ]
+
+jobs:
+
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Set up Go
+      uses: actions/setup-go@v4
+      with:
+        go-version: '1.21'
+
+    - name: Lint
+      uses: golangci/golangci-lint-action@v3
+      with:
+        version: v1.56.0
+
+    - name: Checkstyle 
+      run: make checkstyle 
+
+    - name: Build
+      run: make build 
+
+    - name: Test
+      run: make test
