+++ cmd/container-juggler/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sgeisbacher/container-juggler/internal/cmd"
+
+func main() {
+	cmd.Execute()
+}
