+++ internal/cmd/generate.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/sgeisbacher/container-juggler/generation"
+	"github.com/sgeisbacher/container-juggler/internal/generation"
 	"github.com/spf13/cobra"
 )
 
@@ -16,7 +16,6 @@ var generateCmd = &cobra.Command{
 
 [scenario] defaults to all`,
 	Run: func(cmd *cobra.Command, args []string) {
-		initConfig()
 		fmt.Println("running generate command ...")
 		generator := generation.CreateGenerator()
 		scenario := ""
