+++ internal/cmd/init.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"log"
 
-	"github.com/sgeisbacher/container-juggler/volumeadmin"
+	"github.com/sgeisbacher/container-juggler/internal/volumeadmin"
 	"github.com/spf13/cobra"
 )
 
@@ -12,7 +12,6 @@ var initCmd = &cobra.Command{
 	Short: "initializes all volumes (based on 'volume-init'-config)",
 	Long:  `'init' downloads the zips and extracts them based on 'volume-init'-configuration in container-juggler.yml`,
 	Run: func(cmd *cobra.Command, args []string) {
-		initConfig()
 		loader := volumeadmin.New()
 		if err := loader.Load(false); err != nil {
 			log.Fatal(err)
