+++ internal/cmd/list.go
@@ -11,7 +11,6 @@ var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "lists all available environments",
 	Run: func(cmd *cobra.Command, args []string) {
-		initConfig()
 		scenarios := viper.GetStringMapStringSlice("scenarios")
 		for key := range scenarios {
 			fmt.Println(key)
