+++ internal/cmd/root.go
@@ -28,6 +28,7 @@ func Execute() {
 }
 
 func init() {
+	cobra.OnInitialize(initConfig)
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is ./container-juggler.yml)")
 }
