+++ internal/generation/file_helper.go
@@ -1,7 +1,6 @@
 package generation
 
 import (
-	"io/ioutil"
 	"os"
 )
 
@@ -15,9 +14,9 @@ func (fh DefaultFileHelper) Exists(path string) bool {
 }
 
 func (fh DefaultFileHelper) Write(path string, data string) error {
-	return ioutil.WriteFile(path, []byte(data), 0644)
+	return os.WriteFile(path, []byte(data), 0644)
 }
 
 func (fh DefaultFileHelper) Read(path string) ([]byte, error) {
-	return ioutil.ReadFile(path)
+	return os.ReadFile(path)
 }
