+++ internal/generation/ip_detector.go
@@ -5,15 +5,16 @@ import (
 	"net"
 )
 
+// UplinkIPDetector defines a basic struct
 type UplinkIPDetector struct{}
 
 // Detect detects local-outbound ip-address
 func (ipd UplinkIPDetector) Detect() net.IP {
 	conn, err := net.Dial("udp", "8.8.8.8:53")
-	defer conn.Close()
 	if err != nil {
 		log.Fatal(err)
 	}
+	defer conn.Close()
 
 	switch a := conn.LocalAddr().(type) {
 	case *net.IPAddr:
