--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/sgeisbacher/container-juggler/cmd"
-
-func main() {
-	cmd.Execute()
-}
